% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genlasso_path_ftn.R
\name{ESPgenlasso}
\alias{ESPgenlasso}
\title{Get Exact Solution Path for genlasso problem.}
\usage{
ESPgenlasso(
  y,
  X,
  D,
  genlasso.option = F,
  thres.lambda = 0.1,
  iter = 1e+05,
  tol = 1e-12
)
}
\arguments{
\item{y}{a numeric response vector.}

\item{X}{a matrix of predictor variables.}

\item{D}{a penalty matrix.}

\item{genlasso.option}{whether to use the genlasso package.}

\item{thres.lambda}{minimum threshold lambda.}

\item{iter}{maximum threshold iteration number.}

\item{tol}{tolerance.}
}
\value{
solution path object.
}
\description{
This function solves the genlasso problem for all lambda. So, we can get the solution path object which contains dual solution u, break point lambda, primal solution beta and so on.
}
\examples{
y <- matrix(c(1000), nrow = 1)
X <- matrix(c(1,1,0),nrow = 1)
D <- matrix(c(1,0,0,0,1,1,0,0,-1), nrow = 3)
ESPgenlasso(y,X,D,genlasso.option = FALSE)
}
