% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.beta.R
\name{get.beta}
\alias{get.beta}
\title{Get the primal solution of the genlasso problem at a fixed lambda.}
\usage{
get.beta(object, lambda, tol = 1e-10)
}
\arguments{
\item{object}{ESPgenlasso object.}

\item{lambda}{a specific lambda.}

\item{tol}{tolerance.}
}
\value{
the primal solution at lamdba.
}
\description{
This function gets the primal solution of the genlasso problem at a fixed lambda.
}
\examples{
y <- matrix(c(1000), nrow = 1)
X <- matrix(c(1,1,0),nrow = 1)
D <- matrix(c(1,0,0,0,1,1,0,0,-1), nrow = 3)
object <- ESPgenlasso(y,X,D,genlasso.option = FALSE)
get.beta(object, 700)
}
