% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.abc.R
\name{make.abc}
\alias{make.abc}
\title{Make the parameters in order to find a specific solution for a fixed lambda.}
\usage{
make.abc(object, lambda, tol = 1e-10)
}
\arguments{
\item{object}{ESPgenlasso object.}

\item{lambda}{a specific lambda}

\item{tol}{tolerance.}
}
\value{
parameters A, B and C for the characterization.
}
\description{
In order to find a specific solution, we need some parameters A,B and C which are explained in the paper. This function makes the parameters for a fixed lambda.
}
