% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.linf.solution.R
\name{spec.linf.solution}
\alias{spec.linf.solution}
\title{Get the solution that is a (partial) maximal l infitiy norm solution for some lambda.}
\usage{
spec.linf.solution(object, lambda.seq, max.indices = c(), tol = 1e-10)
}
\arguments{
\item{object}{ESPgenlasso object.}

\item{lambda.seq}{lambda seq.}

\item{max.indices}{beta coef indices that would expected to have the maximum l infinity norm.}

\item{tol}{tolerance.}
}
\value{
solutions for each lambda seq.
}
\description{
This function gets the solution that is a (partial) maximal l infitiy norm solution for some lambda.
}
